# Speed up `TransactionSpec.signTransaction`.